Add files via upload
All assignment specifications were followed for the most part. When user loses a game, it does not restart or quit unless user specifies with commands. Also, refueling and crashes do not happen based on location, commands need to be specified by the player. Entering a blank command repeatedly will give an error. Print display command is implemented in the game world but forgot to add it to game class .